es ar sheimchniot

+++ FlexibleJobs/src/main/java/accounts/Employer.java
@@ -83,7 +83,7 @@ public class Employer implements Account{
     }
 
     @Override
-    public void updateRating(double rate) {
+    public void updateRating(int rate) {
 
     }
